src/main/java/HashMap/FindAllDuplicatesInAnArray.java
package HashMap;

import java.util.ArrayList;
import java.util.HashSet;
import java.util.List;
import java.util.Set;

public class FindAllDuplicatesInAnArray {
    public List<Integer> findDuplicates(int[] nums) {
        List<Integer> answer = new ArrayList<>();
        Set<Integer> set = new HashSet<>();
        for (int num : nums) {
            if (set.contains(num)) {
                answer.add(num);
            } else {
                set.add(num);
            }
        }

        return answer;
    }
}
